bt2: make _create_event_message's parent parameter mandatory
The `parent` parameter of _create_event_message currently has the None
default value.  However, None is never a valid value, as parent must
either be a Stream or a Packet.  Having a None default value makes it
look like the parameter is optional, but is actually always required.

Therefore, remove the None default value from the parameter.

Change-Id: Id0f8d64006ca25281529fc6bb3b9912365799f44
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/1962
Tested-by: jenkins <jenkins@lttng.org>
Reviewed-by: Francis Deslauriers <francis.deslauriers@efficios.com>
Reviewed-by: Philippe Proulx <eeppeliteloop@gmail.com>