ctf: remove unused fields from `babeltrace.trace-info` query
The response of the src.ctf.fs component class to the
`babeltrace.trace-info` query object contains a few fields that are not
used at the moment.  Remove them, before somebody uses them, making us
stuck with them for eternity.

The removed fields are:

- stream class ids
- stream ids
- stream paths
- trace path
- trace range-ns
- trace name

The field `intersection-range-ns` will be removed in a separate patch.

Change-Id: Iee673acbb766cd951e758b5186eabb6fc058b442
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/1990
Tested-by: jenkins <jenkins@lttng.org>
Reviewed-by: Francis Deslauriers <francis.deslauriers@efficios.com>