ctf: remove `intersection-range-ns` from `babeltrace.trace-info` query
The `intersection-range-ns` field of the `src.ctf.fs` component class
response to the `babeltrace.trace-info` query is not essential, since it
can easily be computed from the stream ranges also in the response.
Remove it, and adjust users to compute the intersection themselves.

One user is the CLI and the other is the TraceCollectionMessageIterator
class in the Python bindings.

There is a small implementation detail change in
TraceCollectionMessageIterator.  Previously, we would do one
`babeltrace.trace-info` query per port.  With this patch, we pre-compute
the intersection range for each source component port and store it in a
port name to intersection mapping.  When we create the trimmers, we just look
up that value.

It results in less queries and avoids computing the intersections many
times.  It is also in line with how the CLI does.

Change-Id: Ic5ebded4d810ef992594941dde3316613909c98c
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/1994
Tested-by: jenkins <jenkins@lttng.org>
Reviewed-by: Philippe Proulx <eeppeliteloop@gmail.com>