bt2: test_message_iterator: use assertIs instead of assertTrue/assertFalse in can seek tests
In these tests, we want to specifically check that the return values are
True or False, not any value that evaluates to True or False.

For example, we initialize can_seek_beginning to None, run
graph.run_once(), then do assertFalse(can_seek_beginning).  If, for some
reason, can_seek_beginning is still None, the test will wrongfully pass.

The same could happen if can_seek_beginning gets assigned some value
that evaluates to True (other than True itself) and we use
assertTrue(can_seek_beginning).

Instead, use assertIs, which checks that the value is specifically the
one and only True or the one and only False.

Change-Id: Idf94d350ecc06cdb6a38ca31b2d77804fbf57df7
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/2019
Tested-by: jenkins <jenkins@lttng.org>
Reviewed-by: Francis Deslauriers <francis.deslauriers@efficios.com>
Reviewed-by: Philippe Proulx <eeppeliteloop@gmail.com>