lib: pass config object to message iterator init method, add can seek forward property
This patch introduces a new bt_self_message_iterator_configuration
object that is passed to message iterator init methods.  The purpose of
this object is to hold options the message iterator can set at init time
but not later.

To verify that the user does not save a pointer to this object and try
to use it later, the object is frozen right after the init method
returns.

The object contains one option that we know we'll need,
`can_seek_forward`.  The corresponding functions to set it (on the
config object) and read it (from the self component port input message
iterator object) are also provided.  But because this patch already
contains a lot of boilerplate, the code that actually uses that option
will come in a following patch.

The Python bindings provide a simple wrapper around this new object.

Change-Id: Ibcd0df842fd154d3bafe3869b56f70aa3a72ceab
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/2064
Tested-by: jenkins <jenkins@lttng.org>